+++ src/main/java/io/roastedroot/proxywasm/v1/NetworkContext.java
@@ -16,7 +16,7 @@ public class NetworkContext extends Context {
     }
 
     public Action callOnNewConnection() {
-        int result = proxyWasm.exports().proxyOnNewConnection(id);
+        int result = proxyWasm.abi().proxyOnNewConnection(id);
         Action action = Action.fromInt(result);
         handler.setAction(StreamType.DOWNSTREAM, action);
         return action;
@@ -24,7 +24,7 @@ public class NetworkContext extends Context {
 
     public Action callOnDownstreamData(boolean endOfStream) {
         var data = handler.getDownStreamData();
-        var result = proxyWasm.exports().proxyOnDownstreamData(id, len(data), endOfStream ? 1 : 0);
+        var result = proxyWasm.abi().proxyOnDownstreamData(id, len(data), endOfStream ? 1 : 0);
         Action action = Action.fromInt(result);
         handler.setAction(StreamType.DOWNSTREAM, action);
         return action;
@@ -32,14 +32,14 @@ public class NetworkContext extends Context {
 
     public Action callOnUpstreamData(boolean endOfStream) {
         var data = handler.getUpstreamData();
-        var result = proxyWasm.exports().proxyOnUpstreamData(id, len(data), endOfStream ? 1 : 0);
+        var result = proxyWasm.abi().proxyOnUpstreamData(id, len(data), endOfStream ? 1 : 0);
         Action action = Action.fromInt(result);
         handler.setAction(StreamType.UPSTREAM, action);
         return action;
     }
 
     public void callOnDownstreamConnectionClose(PeerType type) {
-        proxyWasm.exports().proxyOnDownstreamConnectionClose(id, type.getValue());
+        proxyWasm.abi().proxyOnDownstreamConnectionClose(id, type.getValue());
     }
 
     public void callOnDownstreamConnectionClose() {
